requirements.txt
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
et-xmlfile==1.0.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.4.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
pandas==1.1.4
pkg-resources==0.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.11
webdriver-manager==3.2.2
